Prefactoring: Prepare operator for more handlers
