+++ pkg/operator/webhook/validation/add.go
@@ -6,24 +6,17 @@ package validation
 
 import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
-	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
+	"github.com/gardener/gardener/pkg/operator/webhook/validation/garden"
 )
 
-const (
-	// HandlerName is the name of this admission webhook handler.
-	HandlerName = "validator"
-	// WebhookPath is the HTTP handler path for this admission webhook handler.
-	WebhookPath = "/webhooks/validate-operator-gardener-cloud-v1alpha1-garden"
-)
+// AddToManager adds all webhook handlers to the given manager.
+func AddToManager(mgr manager.Manager) error {
+	if err := (&garden.Handler{
+		Logger: mgr.GetLogger().WithName("webhook").WithName(garden.HandlerName),
+	}).AddToManager(mgr); err != nil {
+		return err
+	}
 
-// AddToManager adds Handler to the given manager.
-func (h *Handler) AddToManager(mgr manager.Manager) error {
-	webhook := admission.
-		WithCustomValidator(mgr.GetScheme(), &operatorv1alpha1.Garden{}, h).
-		WithRecoverPanic(true)
-
-	mgr.GetWebhookServer().Register(WebhookPath, webhook)
 	return nil
 }
