+++ pkg/operator/webhook/validation/garden/add.go
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package garden
+
+import (
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
+
+	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
+)
+
+const (
+	// HandlerName is the name of this admission webhook handler.
+	HandlerName = "garden-validator"
+	// WebhookPath is the HTTP handler path for this admission webhook handler.
+	WebhookPath = "/webhooks/validate-operator-gardener-cloud-v1alpha1-garden"
+)
+
+// AddToManager adds Handler to the given manager.
+func (h *Handler) AddToManager(mgr manager.Manager) error {
+	if h.RuntimeClient == nil {
+		h.RuntimeClient = mgr.GetClient()
+	}
+
+	webhook := admission.
+		WithCustomValidator(mgr.GetScheme(), &operatorv1alpha1.Garden{}, h).
+		WithRecoverPanic(true)
+
+	mgr.GetWebhookServer().Register(WebhookPath, webhook)
+	return nil
+}
