+++ pkg/operator/webhook/validation/garden/handler.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package validation
+package garden
 
 import (
 	"context"
